# Make vectorization of CONST explicit