--- a/tinygrad/codegen/uopgraph.py
+++ b/tinygrad/codegen/uopgraph.py
@@ -207,7 +207,6 @@ constant_folder = PatternMatcher([
   (NOp(UOps.GEP, src=(NOp.cvar("c"),)).name("root"), lambda root, c: root.const(c.arg)),
   *[(UPat(UOps.GEP, name="root", src=(UPat(UOps.VECTORIZE, name="v", src=[UPat(UOps.CONST) for _ in range(i)]))),
      lambda root, v: root.const(v.src[root.arg].arg)) for i in [2, 4, 8]],
-  (NOp(UOps.GEP, src=(NOp.cvar("c"),)).name("root"), lambda root, c: root.const(c.arg)),
   (UPat(UOps.CAST, name="root", src=UPat(UOps.CONST, name="c")), lambda root, c: root.const(c.arg)),
   # a phi on a DEFINE_ACC without loops, CONST, or a vectorized CONST is a noop. this is for correctness, not just speed
   (NOp(UOps.PHI, src=(NOp(UOps.DEFINE_ACC).name("acc"), NOp.var("acc"))), lambda acc: UOp.cast(acc.src[0], acc.dtype)),