--- a/tinygrad/codegen/uops.py
+++ b/tinygrad/codegen/uops.py
@@ -102,9 +102,11 @@ class UOp:
       if self.arg is BinaryOps.MUL: return any(x.divides(v) for x in self.src)
     return False # generic false if we aren't sure
   @functools.cached_property
-  def vmin(self) -> UOp: return x if (x:=self._min_max[0]) is not None else self.const(dtypes.min(cast(DType, self.dtype))).vmin
+  def vmin(self) -> UOp:
+    return x if (x:=self._min_max[0]) is not None else UOp.const(cast(DType, self.dtype).scalar(), dtypes.min(cast(DType, self.dtype))).vmin
   @functools.cached_property
-  def vmax(self) -> UOp: return x if (x:=self._min_max[1]) is not None else self.const(dtypes.max(cast(DType, self.dtype))).vmax
+  def vmax(self) -> UOp:
+    return x if (x:=self._min_max[1]) is not None else UOp.const(cast(DType, self.dtype).scalar(), dtypes.max(cast(DType, self.dtype))).vmax
   @functools.cached_property
   def _min_max(self) -> Tuple[Optional[UOp], Optional[UOp]]:
     # NOTE: returned UOp is assumed to be CONST
@@ -112,8 +114,7 @@ class UOp:
     # TODO: UOps.SPECIAL is UOps.DEFINE_VAR
     if self.op is UOps.SPECIAL: return self.const(0), self.const(self.arg[1]-1) if isinstance(self.arg[1], int) else None
     if self.op is UOps.CONST: return self, self
-    if self.op is UOps.VECTORIZE and all(x.op == UOps.CONST for x in self.src):
-      return min(self.src, key=lambda x: x.arg), max(self.src, key=lambda x: x.arg)
+    if self.op is UOps.VECTORIZE and all(x.op == UOps.CONST and x.arg == self.src[0].arg for x in self.src): return self.src[0], self.src[0]
     if self.op is UOps.ALU:
       if self.arg is UnaryOps.NEG and self.dtype != dtypes.bool and not dtypes.is_unsigned(cast(DType, self.dtype)):
         return self.const(-self.src[0].vmax.arg).vmin, self.const(-self.src[0].vmin.arg).vmax