--- a/tinygrad/codegen/uops.py
+++ b/tinygrad/codegen/uops.py
@@ -198,8 +198,7 @@ def type_verify(uops):
       if uop is UOps.CONST:
         assert dtype is not None and dtype == dtype.scalar(), f"consts should be scalar, got {dtype}"
         assert type(arg) is type(dtypes.as_const(arg, dtype)), f"type of {arg=} does not match {dtype}"
-      if uop is UOps.DEFINE_ACC:
-        assert dtype is not None and src[0].dtype == dtype, f"dtype mismatch {src[0].dtype=} != {dtype=}"
+      if uop is UOps.DEFINE_ACC: assert dtype is not None and src[0].dtype == dtype, f"dtype mismatch {src[0].dtype=} != {dtype=}"
     if uop in {UOps.CAST, UOps.BITCAST, UOps.VECTORIZE}: assert arg is None and dtype is not None # type is the output type, not an arg
     if uop is UOps.CAST: assert dtype.count == 1 and len(src) == 1
     if uop is UOps.VECTORIZE: