… test/test_uop_graph.py
… tinygrad/renderer/assembly.py
… tinygrad/renderer/cstyle.py