Resolves #5313