@greg-niemeyer 2024-07-08
@Qazalin regarding

> Should also fix DEFINE_ACC's extra type_verify assert.

I was looking at [you convo with geohot regarding this](https://discord.com/channels/1068976834382925865/1068982781490757652/1254860202566488165 )but I wasn't able to come to a firm conclusion whether you were saying that fixing #5313 requires an update to the assert, or will be implicitly fixed by the requested change. If the assert needs updating, please let me know.