@greg-niemeyer 2024-07-08
I'm diff'ing process replay runs + ASSERT_PROCESS_REPLAY=0 with and without commit 61f9453 and I'm seeing different kernels getting generated. So looks like I've got more work to do.