@greg-niemeyer 2024-07-09
Note: I added the GEP-VECTORIZE-CONST rule because I was getting some strange UOps sequences without it after updating UOps.const(). Attaching GRAPHUOPS=1 images with (gg.uops.svg) and without the rule (bg.uops.svg) -- note the extra PHI-GEP-VECTORIZE-CONST on the bottom of bg.uops.svg.

Adding `assert dtype.count == 1` to render_const causes a bunch of tests to fail, so I'll investigate this and see where that takes me. I'm guessing the reason I'm failing the assert is connected to the CI failures I'm seeing.

![gg uops](https://github.com/tinygrad/tinygrad/assets/152219575/ba340a69-f45e-4619-bc1d-3fd7729ec814)
![bg uops](https://github.com/tinygrad/tinygrad/assets/152219575/1a45e873-e16a-45bb-99b6-9ada4627d491)