@greg-niemeyer 2024-07-09
> @greg-niemeyer when I was playing around with the code, when rendering DEFINE_ACC, it calls render_const with vectorized dtypes sometimes. That might be part of the assert fail.

right, I'm digging into the DEFINE_ACC stuff as we speak