@greg-niemeyer 2024-07-09
Had to add [the render_cast for !(int, bool, float)](https://github.com/tinygrad/tinygrad/pull/5322/commits/499f7e85ab764e90dd7a000ce5ed0b7d14393c52) back -- there are longs, shorts, etc. that need their long-form cast rendered.