@greg-niemeyer 2024-07-09
I'm not super happy with [DEFINE_ACC implicitly vectorizing CONST in cstyle](https://github.com/tinygrad/tinygrad/pull/5322/files#diff-9694d15708a36a57a7e1485db60b1b4b1d2fc1e7278bc87127a73efd0bfa7dc0R171-R173). Should I update DEFINE_ACC in UOps such that its src is VECTORIZE or CONST, consistent with its dtype?