@greg-niemeyer 2024-07-10
Also, I'm pretty sure the UPats I wrote are nonsensical and need to be redone. I will be busy the next couple of days but I'll work on this when I can.