@greg-niemeyer 2024-07-11
> Also, I'm pretty sure the UPats I wrote are nonsensical and need to be redone. I will be busy the next couple of days but I'll work on this when I can.

i rewrote the UPats, but they're very verbose. i plan to rework them after i figure out this uop issue.