@greg-niemeyer 2024-07-13
I'm able to reproduce the GPU/IMAGE and openpilot test failures locally on my M1 macbook with only these two commits:
0152a86c66ff7d7432d5c6fc8ecdf4063ff03155: remove test_const_vectorize_fold 
0ff3cb92c13e291b216d949c5a48015f213873a1: remove const folding UPat for VECTORIZE

As I understand it, these patches shouldn't result in test failures, only process_replay failures due to extra casts. Probing further, I found that these patches pass the same tests prior to migrating to the new lowerer, but fail afterward.
[This branch](https://github.com/greg-niemeyer/tinygrad/tree/pre-lowerer) was synced to the commit immediately preceding the lowerer change, while [this one](https://github.com/greg-niemeyer/tinygrad/tree/post-lowerer) was synced to the lowerer change itself prior to cherry-picking the aforementioned commits.

The test commands I'm using are:
- `GPU=1 IMAGE=2 python -m pytest -n=auto test/test_ops.py`
- `PYTHONPATH="." FLOAT16=0 DEBUGCL=1 GPU=1 IMAGE=2 python examples/openpilot/compile2.py`
- `JIT=2 METAL=1 python -m pytest -n=auto test/ --ignore=test/external --ignore=test/models --durations=20`
  - The METAL tests should pass regardless of whether my two patches are present.

`FORWARD_ONLY=1 GPU=1 IMAGE=2 python -m pytest test/test_ops.py::TestOps::test_simple_padding_conv2d -s` is what I use for most of my debugging, since it seems to be the simplest repro case I have.

Two data points I think are interesting:
- I was able to get the tests to pass again by updating the fix_image_idx() function using CONSTs instead of VECTORIZE in the LOAD. Demo branch [here](https://github.com/greg-niemeyer/tinygrad/tree/post-lowerer-fix-image-idx-hack). Perhaps it's obvious that this would "fix" the issue given the error signature, but the confirmation is nice.
- I was able to get the GPU/IMAGE tests by commenting out a UPat that folds CONSTs via exec_alu(). Demo branch [here](https://github.com/greg-niemeyer/tinygrad/tree/post-lowerer-with-ALU-CONST-hack).
  - This one was a bit of a surprise to me, though perhaps it should be taken with a grain of salt since it also caused a bunch of the METAL tests to fail.

I tried a bunch of other experiments printing and stepping through various parts of `cstyle`, and `uopgraph` with the debugger, but I'm having trouble narrowing the issue down furtherer. Any suggestions would be appreciated.