@greg-niemeyer 2024-07-14
I instrumented the PatternMatcher rules and identified [this UPat](https://github.com/tinygrad/tinygrad/pull/5322/files#diff-00bd44b667ec90ae1d3e984e699bc6b498c84ca1b1bd15a025437ded227457bfR500) as the source of the divergent behavior when using VECTORIZE instead of CONST in [the post-lowerer branch](https://github.com/greg-niemeyer/tinygrad/tree/post-lowerer). Changing the rule to use UOp.var() instead of UOp.cvar() results in VECTORIZE being matched properly and fixes the GPU/IMAGE and openpilot tests. The AMD test is still failing, so that's the next thing to work on.