@Qazalin 2024-07-14
can you use vectorize for WMMA too?