@greg-niemeyer 2024-07-15
Didn't have too much time to work on this today, but I managed to fix the AMD test failure and address some of the review feedback. Tomorrow I'll have more time, so I should be able to make decent progress on cleaning up the UPats, writing tests, and WMMA.