@greg-niemeyer 2024-07-16
> can you use vectorize for WMMA too?

I dumped the UOps.WMMAs produced when running test_linearizer.py and they look like they're getting vectorized already -- they all look like:
UOps.WMMA           : dtypes._half2             [<UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>] ('WMMA_8_8_8_half_half', (8, 8, 8), dtypes.half, dtypes.half, (2, 2, 2), 'METAL', (9, 9, 9), (8,))
UOps.WMMA           : dtypes._float2            [<UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>] ('WMMA_8_8_8_float_float', (8, 8, 8), dtypes.float, dtypes.float, (2, 2, 2), 'METAL', (9, 9, 9), (8,))
UOps.WMMA           : dtypes._float2            [<UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>, <UOps.VECTORIZE: 16>] ('WMMA_8_8_8_half_float', (8, 8, 8), dtypes.half, dtypes.float, (2, 2, 2), 'METAL', (10, 10, 10), (9,))

Did you have something else in mind?

Also, I'll look into what I can do to fix process_replay