@greg-niemeyer 2024-08-01
> holding on merging this because AMD got a big performance regression. I think their HIP compiler is missing optimizations it would've done otherwise.

this perf regression was completely my fault. I wrote some pathological UPats that used lists instead of tuple for VECTORIZE.src. since we are using vec(8)'s on AMD, we were doing 8! (+40,000) comparisons when we should have been doing one, worst-case.

still seeing a minor perf regression. with my change, sdxl takes roughly 800ms longer on the first step, then is 300ms faster on the second step, so ~500ms slower in aggregate on AMD=1 sdxl.
this branch:
```
(tinygrad) greg-niemeyer@tiny10:~/tinygrad$ PYTHONPATH="." AMD=1 python3 examples/sdxl.py --seed 0 --noshow --timing --weights=./weights/sd_xl_base_1.0.safetensors --out=./rendered.png | tee sdxl_good_enuff.txt
ram used:  6.93 GB, sigmas                                            : 100%|█|
100%|███████████████████████████████████████████| 10/10 [00:26<00:00,  0.38it/s]
loaded weights in 17849.89 ms, 6.93 GB loaded at 0.39 GB/s
created batch
step in 13785.81 ms, using 5.30 GB
step in 3052.05 ms, using 7.59 GB
step in 1291.28 ms, using 7.59 GB
step in 1161.24 ms, using 7.59 GB
step in 1162.83 ms, using 7.59 GB
step in 1155.86 ms, using 7.59 GB
step in 1162.48 ms, using 7.59 GB
step in 1157.88 ms, using 7.59 GB
step in 1166.16 ms, using 7.59 GB
step in 1151.83 ms, using 7.59 GB
```

master:
```
(tinygrad) greg-niemeyer@tiny10:~/tinygrad$ PYTHONPATH="." AMD=1 python3 examples/sdxl.py --seed 0 --noshow --timing --weights=./weights/sd_xl_base_1.0.safetensors --out=./rendered.png
| tee sdxl_master.txt
ram used:  6.93 GB, sigmas                                            : 100%|█|
100%|███████████████████████████████████████████| 10/10 [00:25<00:00,  0.39it/s]
loaded weights in 17865.05 ms, 6.93 GB loaded at 0.39 GB/s
created batch
step in 13015.36 ms, using 5.30 GB
step in 3394.70 ms, using 7.59 GB
step in 1295.29 ms, using 7.59 GB
step in 1163.01 ms, using 7.59 GB
step in 1163.04 ms, using 7.59 GB
step in 1165.24 ms, using 7.59 GB
step in 1166.17 ms, using 7.59 GB
step in 1165.33 ms, using 7.59 GB
step in 1167.06 ms, using 7.59 GB
step in 1158.80 ms, using 7.59 GB
```